+++ src/style/DevLink.less
@@ -1,0 +1,18 @@
+.dev-link {
+    background-color: #ffd3ba;
+    border-radius: 10px;
+    display: inline-flex;
+    margin: 20px 0px;
+    box-shadow: -2px 2px 2px rgba(0, 0, 0, 0.2);
+    text-decoration: none !important;
+    align-self: start;
+    .dev-label {
+        background-color: #ffad14;
+        border-right: 1px solid black;
+        border-bottom-left-radius: 10px;
+        border-top-left-radius: 10px;
+    }
+    .dev-label,.dev-text {
+        padding: 5px 10px;
+    }
+}
